# go-app/docs/api/history/openapi.yaml
openapi: 3.0.3
info:
  title: Alert History API
  description: |
    Comprehensive API for querying alert history with advanced filtering, pagination, and analytics.

    ## Features
    - 18+ filter types (status, severity, namespace, labels, time range, etc.)
    - Pagination and sorting
    - 2-tier caching (L1 in-memory + L2 Redis)
    - Performance optimized (p95 < 10ms)
    - OWASP Top 10 compliant

    ## Authentication
    All endpoints require API key authentication via `X-API-Key` header.
  version: 2.0.0
  contact:
    name: Alert History API Support
    email: support@alert-history.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.alert-history.example.com/api/v2
    description: Production server
  - url: https://staging-api.alert-history.example.com/api/v2
    description: Staging server
  - url: http://localhost:8080/api/v2
    description: Local development server
...
paths:
...
  /history/{fingerprint}:
    get:
      tags:
        - History
      summary: Get alert timeline
      description: Retrieve all alerts with the same fingerprint (alert timeline)
      operationId: getAlertTimeline
      parameters:
        - name: fingerprint
          in: path
          required: true
          description: Alert fingerprint (64 hex characters)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{64}$'
          example: "a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef1234567890"
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          example: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  fingerprint:
                    type: string
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  count:
                    type: integer
                  limit:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    Alert:
      type: object
      required:
        - fingerprint
        - alert_name
        - status
        - starts_at
      properties:
        fingerprint:
          type: string
          description: Alert fingerprint (64 hex characters)
          example: "a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef1234567890"
        alert_name:
          type: string
          description: Alert name
          example: "HighCPUUsage"
        status:
          type: string
          enum: [firing, resolved]
          example: "firing"
        severity:
          type: string
          enum: [critical, warning, info, noise]
          example: "critical"
        namespace:
          type: string
          example: "production"
        labels:
          type: object
          additionalProperties:
            type: string
          example:
            instance: "server-01"
            job: "node-exporter"
        annotations:
          type: object
          additionalProperties:
            type: string
        starts_at:
          type: string
          format: date-time
          example: "2025-11-16T10:00:00Z"
        ends_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-16T11:00:00Z"
        generator_url:
          type: string
          format: uri
          nullable: true
        timestamp:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid query parameter"
            details:
              type: object
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time
            documentation_url:
              type: string
              format: uri
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid query parameter: page must be between 1 and 10000"
              request_id: "req-12345"
              timestamp: "2025-11-16T12:00:00Z"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...